Remove bash specific statement

+++ rdr/busybox/scripts/kernel_untar.sh
@@ -1,14 +1,16 @@
-#!/bin/bash
-
-KERNEL_ZIP_URL="https://github.com/torvalds/linux/archive/refs/heads/master.zip"
+KERNEL_TAR_URL="https://cdn.kernel.org/pub/linux/kernel/v4.x/linux-4.14.280.tar.gz"
+KERNEL_VERSION=`echo $KERNEL_TAR_URL | cut -d '/' -f8 | sed 's/.tar.gz$//'`
 MOUNT=/mnt/test/
-MASTER_COPY=master_kernel.zip
-KERNEL_DIRECTORY="linux-master"
+MASTER_COPY=$KERNEL_VERSION".tar.gz"
+KERNEL_DIRECTORY=$KERNEL_VERSION
 MASTER_CHECKSUM_FILE="$MOUNT/kernel_hash"
 # Allow only upto 90% full disk space
-DISK_SPACE_FULL_THRESHOLD=90
+DISK_SPACE_FULL_THRESHOLD=70
 # During make free space, don't delete below 50% usage
 FREE_SPACE_THRESHOLD=50
+# SIGNALLING THROUGH FILES
+PAUSE_FILE="IO_PAUSE"
+RUNNING_FILE="IO_RUNNING"
 
 get_disk_usage()
 {
@@ -23,11 +25,16 @@ get_disk_usage()
 # beginning directory creation we will still have some data
 make_free_space()
 {
-    testdir=`ls -d linux-master_[0-9]* | sed -n 1p`
-    rm -rf $testdir
-    cur_usage=$(get_disk_usage)
-    if [ $cur_usage -gt $FREE_SPACE_THRESHOLD ]; then
-        make_free_space
+    ls -d "$KERNEL_VERSION"_[0-9]*
+    if [ $? -eq 0 ]; then
+        testdir=`ls -d "$KERNEL_VERSION"_[0-9]* | sed -n 1p`
+        echo "Cleaning up $testdir "
+        rm -rf $testdir
+        sync -f $MOUNT
+        cur_usage=$(get_disk_usage)
+        if [ $cur_usage -gt $FREE_SPACE_THRESHOLD ]; then
+            make_free_space
+        fi
     fi
 }
 
@@ -42,31 +49,56 @@ block_if_no_space_left()
     fi
 }
 
-if [ -f $MASTER_COPY ]; then
-    echo "Master copy zip already exists"
+# IO status needs to be paused if checksum needs to be calculated
+wait_if_pause()
+{
+    while [ -f $PAUSE_FILE ]
+    do
+        if [ -f $RUNNING_FILE ]; then
+            unlink $RUNNING_FILE
+            sync -f $MOUNT
+        fi
+        sleep 10
+    done
+}
+
+# set running state
+set_running_state()
+{
+    touch $RUNNING_FILE
+    sync -f $MOUNT
+}
+
+if [ -f "$MOUNT""/""$MASTER_COPY" ]; then
+    echo "Master copy tar already exists"
 else
-    echo "Downloading kernel zip"
-    wget $KERNEL_ZIP_URL -O $MOUNT/$MASTER_COPY
+    echo "Downloading kernel tar"
+    wget $KERNEL_TAR_URL -O $MOUNT/$MASTER_COPY
     if [ $? -eq 0 ];
     then
-        echo "Kernel zip downloaded successfully"
+        echo "Kernel tar ball downloaded successfully"
     else
         echo "Failed to download kernel"
         exit 1
     fi
 fi
 
-if [ -f "$MOUNT/$MASTER_COPY" ]; then
-    if [ -d "$MOUNT/$KERNEL_DIRECTORY"_original ]; then
+if [ -f "$MOUNT""/""$MASTER_COPY" ]; then
+    if [ -d "$MOUNT""/""$KERNEL_DIRECTORY"_original ]; then
         echo "Reference Kernel dir already exists"
     else
-        unzip -q $MOUNT/$MASTER_COPY -d $MOUNT/
+        # If we have broken linux-<version> dir, lets clean it up
+        if [ -d "$MOUNT""/""$KERNEL_DIRECTORY" ]; then
+            rm -rf $MOUNT/$KERNEL_DIRECTORY
+        fi
+        cd $MOUNT
+        tar xfz $MASTER_COPY
         if [ $? -ne 0 ];
         then
-            echo "Failed to unzip reference kernel"
+            echo "Failed to untar reference kernel"
             exit 1
         fi
-        arequal-checksum $MOUNT/$KERNEL_DIRECTORY>>$MASTER_CHECKSUM_FILE
+        arequal-checksum $MOUNT/$KERNEL_DIRECTORY>$MASTER_CHECKSUM_FILE
         mv $MOUNT/$KERNEL_DIRECTORY $MOUNT/$KERNEL_DIRECTORY"_original"
     fi
 fi
@@ -74,16 +106,22 @@ fi
 # From here keep a loop of untaring and renaming the kernel dirs
 while true
 do
+    sync -f $MOUNT
     block_if_no_space_left
-    # There could be half unzipped linx-master dir
+    # There could be half untared linx dir
     # may be due to failover , so we need to cleanup
     if [ -d $KERNEL_DIRECTORY ]; then
         rm -rf $KERNEL_DIRECTORY
     fi
-    unzip -o -q $MOUNT/$MASTER_COPY -d $MOUNT/
+    cd $MOUNT
+    wait_if_pause
+    # We are out of pause state
+    # create a file for IO running
+    set_running_state
+    tar xfz $MASTER_COPY
     if [ $? -ne 0 ];
     then
-        echo "Failed to unzip kernel"
+        echo "Failed to untar kernel"
         exit 1
     fi
     mv $MOUNT/$KERNEL_DIRECTORY $MOUNT/$KERNEL_DIRECTORY"_`date +%s`"
